+++ src/environment/Debris.py
@@ -2,7 +2,7 @@
 https://github.com/JerrettYang/Iridium-33-17126UTC
 """
 import numpy as np
-from InPlaneEquations import G , M
+from src.environment.InPlaneEquations import G , M
 from astropy import units as u
